+++ BUS/LoaiPhongBUS.cs
@@ -15,7 +15,7 @@ namespace BUS
         {
             return lpdal.getalllp();
         }
-        public decimal donGia(string maLoaiPhong)
+        public double donGia(string maLoaiPhong)
         {
             return lpdal.donGia(maLoaiPhong);
         }
